LIBFCREPO-1189. Restructured the plastron.jobs package.
https://umd-dit.atlassian.net/browse/LIBFCREPO-1189
