+++ .gitignore
@@ -4,7 +4,7 @@
 build
 config
 dist
-jobs
+/jobs/
 logs
 
 .coverage
